Modify Codes
The output you received indicates that there are many files and folders that have not yet been tracked by Git, i.e., they have not been added to the staging area to be included in the commit. Some of these files and folders may be OS-related (such as the AppData/, Documents/, Downloads/, NTUSER.DAT, etc.) or program-specific (such as VirtualBox VMs/, .vscode/, .android/, etc.) that you probably do not want to include in your Git repository.